requirements.txt
gradio==4.44.1
httpx[http2] 
pydantic==2.10.6
ollama
gunicorn
uvicorn
python-dotenv
beautifulsoup4
